Added balls and collision to balls. Also walls.
Further collision specification probably needs to be made, right now the
collision is only on Balls (but they might be the only ones who need it
so Im on the edge in terms of whatdo).

Also, they need a more randomized 'bounce' , or atleast when they hit
certain elements, or entities- i.e the player paddles, as I'd like
collision off of entities to be different.
Collision can probably be optimized. If a ball is no where near a wall
it probably doesn't need to try and check it's collision, but at the
moment there is no point to optimizing.

Any way next update I'd like to add in player paddles and/or game
sequencing- i.e start up game, main menui, click to start, and there are
other UI stuff that can be layered and will work beautifully thanks to
my the lovely event system. /long commit