Updated dimensions of the paddles, more like pong
Collision is still really wacky because Im not doing it in a smart way,
for the paddle collision. Also wall collision breaks at higher speeds
but this could be a symptom of trigonometry but I doubt it since Im
pretty sure my math is solid. Any way, paddle collision adds more
collisions when any parts overlap, but when I go to calculate a new
angle it only looks at the ball's position.