Did much of the TODO, added sounds
Sounds are disabled right now because the process is spawning a billion
threads so I need to create a sound managing class that will close each
thread in the appropriate time, probably by storing each clip in a list
and checking each sound if it's done playing. It will probably do this
by extending clip to include a lifetime monitor, storing this extending
class in the list and comparing the lifetime monitor against current
time and checking if that is longer than the clip length.